requirements.txt
numpy==1.24.2
opencv-python==4.7.0.72
Pillow==9.5.0
tqdm==4.65.0
scipy~=1.12.0
matplotlib~=3.8.2